+++ packages/commonwealth/client/scripts/views/components/collapsible_body_text.tsx
@@ -1,0 +1,122 @@
+/* @jsx m */
+
+import m from 'mithril';
+
+import app from 'state';
+import { Account, Comment, Thread, AnyProposal } from 'models';
+import { countLinesQuill, countLinesMarkdown } from './quill/helpers';
+import { MarkdownFormattedText } from './quill/markdown_formatted_text';
+import { QuillFormattedText } from './quill/quill_formatted_text';
+import User from './widgets/user';
+
+const QUILL_PROPOSAL_LINES_CUTOFF_LENGTH = 50;
+const MARKDOWN_PROPOSAL_LINES_CUTOFF_LENGTH = 70;
+
+export const formatBody = (vnode, updateCollapse) => {
+  const { item } = vnode.attrs;
+  if (!item) return;
+
+  const body =
+    item instanceof Comment
+      ? item.text
+      : item instanceof Thread
+      ? item.body
+      : item.description;
+  if (!body) return;
+
+  vnode.state.body = body;
+  if (updateCollapse) {
+    try {
+      const doc = JSON.parse(body);
+      if (countLinesQuill(doc.ops) > QUILL_PROPOSAL_LINES_CUTOFF_LENGTH) {
+        vnode.state.collapsed = true;
+      }
+    } catch (e) {
+      if (countLinesMarkdown(body) > MARKDOWN_PROPOSAL_LINES_CUTOFF_LENGTH) {
+        vnode.state.collapsed = true;
+      }
+    }
+  }
+};
+
+export class CollapsibleBodyText
+  implements
+    m.ClassComponent<{
+      item: AnyProposal | Thread | Comment<any>;
+    }>
+{
+  private body: any;
+  private collapsed: boolean;
+
+  oninit(vnode) {
+    this.collapsed = false;
+    formatBody(vnode, true);
+  }
+
+  onupdate(vnode) {
+    formatBody(vnode, false);
+  }
+
+  view(vnode) {
+    const { body } = this;
+
+    const getPlaceholder = () => {
+      if (!(vnode.attrs.item instanceof Thread)) return;
+
+      const author: Account = app.chain
+        ? app.chain.accounts.get(vnode.attrs.item.author)
+        : null;
+
+      return author ? (
+        <>
+          {m(User, {
+            user: author,
+            hideAvatar: true,
+            hideIdentityIcon: true,
+          })}{' '}
+          created this thread
+        </>
+      ) : (
+        'Created this thread'
+      );
+    };
+
+    const text = () => {
+      try {
+        const doc = JSON.parse(body);
+
+        if (!doc.ops) throw new Error();
+
+        if (
+          doc.ops.length === 1 &&
+          doc.ops[0] &&
+          typeof doc.ops[0].insert === 'string' &&
+          doc.ops[0].insert.trim() === ''
+        ) {
+          return getPlaceholder();
+        }
+
+        return (
+          <QuillFormattedText
+            doc={doc}
+            cutoffLines={QUILL_PROPOSAL_LINES_CUTOFF_LENGTH}
+            collapse={false}
+            hideFormatting={false}
+          />
+        );
+      } catch (e) {
+        if (body?.toString().trim() === '') {
+          return getPlaceholder();
+        }
+        return (
+          <MarkdownFormattedText
+            doc={body}
+            cutoffLines={MARKDOWN_PROPOSAL_LINES_CUTOFF_LENGTH}
+          />
+        );
+      }
+    };
+
+    return <>{text()}</>;
+  }
+}
