+++ packages/commonwealth/client/scripts/views/components/comments/helpers.ts
@@ -1,3 +1,6 @@
+import app from 'state';
+import { ContentType } from 'types';
+
 // highlight the header/body of a parent thread, or the body of a comment
 export const jumpHighlightComment = (commentId: number) => {
   const commentEle = document.getElementsByClassName(`comment-${commentId}`)[1];
@@ -14,4 +17,13 @@ export const jumpHighlightComment = (commentId: number) => {
       commentEle.classList.add('highlightAnimationComplete');
     }, 2000 + 500);
   }
+};
+
+export const clearEditingLocalStorage = (
+  id: number | string,
+  contentType: ContentType
+) => {
+  localStorage.removeItem(
+    `${app.activeChainId()}-edit-${contentType}-${id}-storedText`
+  );
 };
