--- packages/commonwealth/client/scripts/views/pages/view_proposal/constants.ts
@@ -1,3 +1,0 @@
-export const QUILL_PROPOSAL_LINES_CUTOFF_LENGTH = 50;
-export const MARKDOWN_PROPOSAL_LINES_CUTOFF_LENGTH = 70;
-export const MAX_THREAD_LEVEL = 2;
